# src/DotNetCqrsApi.Api/openapi.json
{
...
  "swagger": "2.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/modules": {
      "post": {
        "tags": [
          "Modules"
        ],
        "operationId": "Modules_CreateModule",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateModuleRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ModuleModel"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Modules"
        ],
        "operationId": "Modules_UpdateModule",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateModuleRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/operation-businesses": {
      "get": {
        "tags": [
          "OperationBusinesses"
        ],
        "operationId": "OperationBusinesses_GetOperationBusiness",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/OperationBusinessModel"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "OperationBusinesses"
        ],
        "operationId": "OperationBusinesses_CreateOperationBusiness",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateOperationBusinessRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/OperationBusinessModel"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "OperationBusinesses"
        ],
        "operationId": "OperationBusinesses_UpdateOperationBusiness",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateOperationBusinessRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_GetUser",
        "parameters": [
          {
            "type": "integer",
            "name": "id",
            "in": "query",
            "format": "int32",
            "x-nullable": false
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_CreateUser",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CreateUserRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/UserModel"
            }
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_UpdateUser",
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateUserRequest"
            },
            "x-nullable": true
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ValidationProblemDetails"
            }
          },
          "500": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          },
          "502": {
            "x-nullable": false,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ProblemDetails"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "OperationBusinessModel": {
      "type": "object",
      "required": [
        "id",
        "isEnabled",
        "countryId",
        "companyId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "acronym": {
          "type": "string"
        },
        "isEnabled": {
          "type": "boolean"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        },
        "countryName": {
          "type": "string"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        },
        "companyName": {
          "type": "string"
        }
      }
    },
    "ValidationProblemDetails": {
      "allOf": [
        {
          "$ref": "#/definitions/ProblemDetails"
        },
        {
          "type": "object",
          "additionalProperties": {},
          "properties": {
            "errors": {
              "type": "object",
              "additionalProperties": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      ]
    },
    "ProblemDetails": {
      "type": "object",
      "additionalProperties": {},
      "properties": {
        "type": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "detail": {
          "type": "string"
        },
        "instance": {
          "type": "string"
        }
      }
    },
...
    "CountryModel": {
      "type": "object",
      "required": [
        "id",
        "currencyId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "code": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "currencyId": {
          "type": "integer",
          "format": "int32"
        },
        "modules": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ModuleModel"
          }
        }
      }
    },
    "ModuleModel": {
      "type": "object",
      "required": [
        "id",
        "number",
        "countryId",
        "isEnabled"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "number": {
          "type": "number",
          "format": "decimal"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        },
        "country": {
          "type": "string"
        },
        "isEnabled": {
          "type": "boolean"
        }
      }
    },
...
    "CreateModuleRequest": {
      "type": "object",
      "required": [
        "id",
        "number",
        "countryId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "UpdateModuleRequest": {
      "type": "object",
      "required": [
        "id",
        "number",
        "countryId",
        "isEnabled"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "number": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "location": {
          "type": "string"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        },
        "isEnabled": {
          "type": "boolean"
        }
      }
    },
...
    "CreateOperationBusinessRequest": {
      "type": "object",
      "required": [
        "countryId",
        "companyId"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "acronym": {
          "type": "string"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "UpdateOperationBusinessRequest": {
      "type": "object",
      "required": [
        "id",
        "countryId",
        "companyId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "acronym": {
          "type": "string"
        },
        "countryId": {
          "type": "integer",
          "format": "int32"
        },
        "companyId": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "UserModel": {
      "type": "object",
      "required": [
        "id",
        "roleId",
        "isEnabled"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "roleId": {
          "type": "integer",
          "format": "int32"
        },
        "roleName": {
          "type": "string"
        },
        "isEnabled": {
          "type": "boolean"
        },
        "countries": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CountryModel"
          }
        },
        "sections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SectionModel"
          }
        }
      }
    },
    "SectionModel": {
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "CreateUserRequest": {
      "type": "object",
      "required": [
        "roleId",
        "isEnabled"
      ],
      "properties": {
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "roleId": {
          "type": "integer",
          "format": "int32"
        },
        "isEnabled": {
          "type": "boolean"
        },
        "countryIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "moduleIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "sectionIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
    "UpdateUserRequest": {
      "type": "object",
      "required": [
        "id",
        "roleId"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "email": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "surname": {
          "type": "string"
        },
        "roleId": {
          "type": "integer",
          "format": "int32"
        },
        "countryIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "sectionIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        },
        "moduleIds": {
          "type": "array",
          "items": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    },
...
  }
}